Refactor error types, make backwards compatible
